requirements.txt
aiohttp==3.5.4
aiosmtpd==1.2
alabaster==0.7.12
altgraph==0.17
amqp==2.5.2
astroid==2.2.5
async-timeout==3.0.1
atomicwrites==1.3.0
atpublic==1.0
attrs==18.2.0
awscli==1.16.96
Babel==2.7.0
bandit==1.6.2
beautifulsoup4==4.7.1
billiard==3.6.3.0
bleach==3.1.0
bootstrap4==0.1.0
boto3==1.9.86
botocore==1.12.86
bumpversion==0.5.3
celery==4.4.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
coderunner==0.6
colorama==0.3.7
coverage==4.5.3
coveralls==1.8.1
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2
django-appconf==1.0.2
django-bootstrap-form==3.4
django-bootstrap-forms==0.1
Django-Bootstrap3-Validator==0.3.3
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-classy-tags==0.8.0
django-cms==3.5.3
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-filters==0.2.1
django-formtools==2.1
django-heroku==0.3.1
django-javascript-settings==1.1.2
django-jquery==3.1.0
django-js-asset==1.2.2
django-json==0.27
django-json-settings==0.2
django-jsonview==1.2.0
django-mailgun==0.9.1
django-multiselectfield==0.1.12
django-redis==4.10.0
django-redis-cache==2.0.0
django-rest-framework==0.1.0
django-restframework==0.0.1
django-safedelete==0.5.4
django-sekizai==0.10.0
django-smtp-ssl==1.0
django-static-jquery==2.1.4
django-static-jquery3==3.3.1.1
django-sudo==2.1.0
django-template-tags==1.1.3
django-treebeard==4.3
django-watchman==0.15.0
django-wordpress-rss==0.1.0
djangocms-admin-style==1.3.0
djangocms-owl==0.1.11
djangorestframework==3.9.4
djangorestframework-stubs==1.1.0
doc8==0.8.0
docopt==0.6.2
docutils==0.14
entrypoints==0.3
fanstatic==1.1
feedparser==5.1.2
filelock==3.0.12
flake8==3.7.8
flake8-bandit==2.1.1
flake8-broken-line==0.1.0
flake8-bugbear==19.3.0
flake8-builtins==1.4.1
flake8-comprehensions==2.1.0
flake8-debugger==3.1.0
flake8-fixme==1.1.1
flake8-isort==2.7.0
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-variables-names==0.0.1
flufl.i18n==2.0.1
fn==0.4.3
footballdataorg==0.1.6
future==0.17.1
gain-requests-futures==0.9.7
geometric==0.9.7.1
Geometry==0.0.23
gevent==1.4.0
gevent-websocket==0.10.1
gitdb2==2.0.5
GitPython==2.1.11
gmail==0.6.3
gmail-client==0.0.7.7
greenlet==0.4.15
grequests==0.3.0
gunicorn==20.0.4
httplib2==0.12.0
httpsproxy-urllib2==1.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.18
IntelliCoder==0.5.2
invoke==1.2.0
isort==4.3.21
javascript==2015.32.59
Jinja2==2.10.1
jmespath==0.9.3
jquery.javascript==1.0.0
js.angular==1.1.4
js.angular-ui-bootstrap==0.6.0.post2
js.bootstrap==3.3.4
js.javascript-templates==2.1.0
js.jquery==3.3.1
jsonfield==2.0.2
jsonpath==0.81
kombu==4.6.8
lazy-object-proxy==1.3.1
livescore==2018.0.0
loadtest==0.0.2
lxml==4.3.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==5.0.0
multidict==4.5.2
mysqlclient==1.4.2
networkx==2.4
numpy==1.17.4
oauthlib==3.0.1
optional-django==0.3.0
ordereddict==1.1
packaging==19.0
pbr==5.4.0
pefile==2018.8.8
pep8-naming==0.8.2
pev==0.1.0
Pillow==5.4.1
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.12.0
plumbum==1.6.7
progressbar==2.5
prompt-toolkit==1.0.14
psycopg2==2.8.4
ptvsd==4.2.10
putils==0.2.3
py==1.8.0
py-smtp==1.0.8
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.0
pyelftools==0.25
pyflakes==2.1.1
Pygments==2.4.2
PyInquirer==1.0.3
PyInstaller==3.6
PyIntelliJ2Eclipse==0.1
pylint==2.3.1
PyMySQL==0.9.3
PyMysqlDB==0.0.2
pyparsing==2.4.0
pypi==2.1
pypl==40.10
pytest==5.0.1
pytest-cov==2.7.1
pytest-runner==5.1
python-dateutil==2.7.5
python-decouple==3.1
python-dev-tools==2019.7.9
python-gettext==4.0
python-magic==0.4.15
python-redis-lock==3.3.1
python-wordpress==0.1.3
pytz==2018.9
pywin32-ctypes==0.2.0
PyYAML==3.13
react==4.3.0
readme-renderer==24.0
redis==3.2.1
redis-structures==0.1.7
regex==2020.1.8
requests==2.21.0
requests-cache==0.5.0
requests-futures==1.0.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
restructuredtext-lint==1.3.0
rsa==3.4.2
s3transfer==0.1.13
secure-smtpd==3.0.0
sh==1.12.0
shutilwhich==1.1.0
six==1.12.0
smmap2==2.0.5
smtp-faker==0.1.2
smtp-mail==0.0.2
smtp2go==2.0.0
snowballstemmer==1.9.0
soupsieve==1.7.3
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.2.17
sqlparse==0.2.4
stevedore==1.30.1
sudoaptinstall==0.3
svgwrite==1.3.1
sweetify==2.0.0
testfixtures==6.10.0
toml==0.10.0
tox==3.13.2
tqdm==4.32.2
transaction==2.4.0
twine==1.13.0
typed-ast==1.3.1
urllib2-file==0.2.1
urllib3==1.24.1
urllib5==5.0.0
vine==1.3.0
virtualenv==16.6.1
virtualenv-clone==0.5.3
watchman==0.0.8
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==0.14.1
whitenoise==5.0.1
wordpress-api==1.2.8
wordpress-json==0.3.1
wrapt==1.11.1
wsjson==0.9
xmljson==0.2.0
xmltodict==0.11.0
yarl==1.3.0
z3c.json==0.5.5
zipp==0.5.2
zope.browser==2.3
zope.browserpage==4.3.0
zope.component==4.5
zope.configuration==4.3.0
zope.contentprovider==4.2.1
zope.contenttype==4.4
zope.deferredimport==4.3
zope.deprecation==4.4.0
zope.event==4.4
zope.exceptions==4.3
zope.hookable==4.2.0
zope.i18n==4.6.1
zope.i18nmessageid==4.3.1
zope.interface==4.6.0
zope.location==4.2
zope.pagetemplate==4.4.1
zope.proxy==4.3.1
zope.publisher==5.0.1
zope.schema==4.9.3
zope.security==4.3.1
zope.tal==4.4
zope.tales==4.3
zope.traversing==4.3.1
zope.viewlet==4.2.1
